update.js: Rename _useFlattenMap.js
